+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Dtos/ChargeLinksCommands/Validation/BusinessValidation/ValidationRules/ChargeLinksUpdateNotYetSupportedRule.cs
@@ -14,7 +14,6 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using GreenEnergyHub.Charges.Core.DateTime;
 using GreenEnergyHub.Charges.Domain.ChargeLinks;
 using GreenEnergyHub.Charges.Domain.Dtos.Validation;